+++ tests/unit_tests/commands_test.js
@@ -1,4 +1,5 @@
 import "./test_helper.js";
+import "../../background_scripts/tab_recency.js";
 import "../../background_scripts/bg_utils.js";
 import "../../lib/settings.js";
 import "../../lib/keyboard_utils.js";
@@ -113,6 +114,21 @@ context("parseKeyMappingConfig", () => {
       Commands.parseKeyMappingsConfig("mapkey a b \n unmap a a").keyToMappedKey,
     );
   });
+
+  should("return validation errors", () => {
+    const getErrors = (config) => Commands.parseKeyMappingsConfig(config).validationErrors;
+    assert.equal(0, getErrors("map a scrollDown").length);
+    // Missing an action (map).
+    assert.equal(1, getErrors("a scrollDown").length);
+    // Invalid action.
+    assert.equal(1, getErrors("invalidAction a scrollDown").length);
+    // Unmap allows only 1 argument.
+    assert.equal(0, getErrors("unmap a").length);
+    assert.equal(1, getErrors("unmap a b").length);
+    // Mapkey requires 2 arguments.
+    assert.equal(0, getErrors("mapkey a b").length);
+    assert.equal(1, getErrors("mapkey a").length);
+  });
 });
 
 context("Validate commands and options", () => {
@@ -142,13 +158,6 @@ context("Validate commands and options", () => {
     }
   });
 
-  should("have valid commands for each advanced command", () => {
-    for (const command of Commands.advancedCommands) {
-      assert.equal("string", typeof command);
-      assert.isTrue(Commands.availableCommands[command]);
-    }
-  });
-
   should("have valid commands for each default key mapping", () => {
     const count = Object.keys(Commands.keyToRegistryEntry).length;
     assert.isTrue(0 < count);
@@ -156,16 +165,6 @@ context("Validate commands and options", () => {
       const command = Commands.keyToRegistryEntry[key];
       assert.equal("object", typeof command);
       assert.isTrue(Commands.availableCommands[command.command]);
-    }
-  });
-});
-
-context("Validate advanced commands", () => {
-  should("include each advanced command in a command group", () => {
-    const allCommands = Object.keys(Commands.commandGroups).map((k) => Commands.commandGroups[k])
-      .flat(1);
-    for (const command of Commands.advancedCommands) {
-      assert.isTrue(allCommands.includes(command));
     }
   });
 });