+++ tests/unit_tests/test_helper.js
@@ -17,4 +17,4 @@ const shouldaSubset = {
 globalThis.isUnitTests = true;
 
 // Attach shoulda's functions, like setup, context, should, to the global namespace.
-Object.assign(window, shouldaSubset);
+Object.assign(globalThis, shouldaSubset);